--- FakeNFT/Models/Profile.swift
@@ -1,9 +1,0 @@
-//
-// Created by Сергей Махленко on 04.06.2023.
-//
-
-import Foundation
-
-struct UpdateProfileNfts: Codable {
-    let nfts: [Int]
-}
