--- FakeNFT/Scenes/Cart/PaySuccessful/PaySucessfulViewModel.swift
@@ -1,8 +1,0 @@
-//
-// Created by Сергей Махленко on 03.06.2023.
-//
-
-import Foundation
-
-class PaySucessfulViewModel {
-}
